requirements.txt
click==6.7
Flask==0.12.2
Flask-JWT==0.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
pkg-resources==0.0.0
PyJWT==1.4.2
pymongo==3.6.0
python-dotenv==0.7.1
six==1.11.0
Werkzeug==0.14.1
